move to e2e_test_utils instead of test_utils
