+++ crates/sui-bridge/src/node.rs
@@ -160,9 +160,9 @@ mod tests {
     use crate::config::BridgeNodeConfig;
     use crate::config::EthConfig;
     use crate::config::SuiConfig;
+    use crate::e2e_test_utils::deploy_sol_contract;
+    use crate::e2e_test_utils::get_eth_signer_client_e2e_test_only;
     use crate::e2e_test_utils::wait_for_server_to_be_up;
-    use crate::test_utils::deploy_sol_contract;
-    use crate::test_utils::get_eth_signer_client_e2e_test_only;
     use crate::BRIDGE_ENABLE_PROTOCOL_VERSION;
     use fastcrypto::secp256k1::Secp256k1KeyPair;
     use sui_config::local_ip_utils::get_available_port;
