+++ crates/sui-bridge/src/test_utils.rs
@@ -1,9 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::abi::EthBridgeCommittee;
 use crate::abi::EthToSuiTokenBridgeV1;
-use crate::crypto::BridgeAuthorityPublicKeyBytes;
 use crate::eth_mock_provider::EthMockProvider;
 use crate::events::SuiBridgeEvent;
 use crate::server::mock_handler::run_mock_server;
@@ -11,8 +9,6 @@ use crate::sui_transaction_builder::build_sui_transaction;
 use crate::types::{
     BridgeCommitteeValiditySignInfo, CertifiedBridgeAction, VerifiedCertifiedBridgeAction,
 };
-use crate::utils::get_eth_signer_client;
-use crate::utils::EthSigner;
 use crate::{
     crypto::{BridgeAuthorityKeyPair, BridgeAuthorityPublicKey, BridgeAuthoritySignInfo},
     events::EmittedSuiToEthTokenBridgeV1,
@@ -32,15 +28,10 @@ use fastcrypto::encoding::{Encoding, Hex};
 use fastcrypto::traits::KeyPair;
 use hex_literal::hex;
 use move_core_types::language_storage::TypeTag;
-use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, HashMap};
-use std::fs::File;
-use std::io::{Read, Write};
 use std::net::IpAddr;
 use std::net::Ipv4Addr;
 use std::net::SocketAddr;
-use std::process::Command;
-use std::str::FromStr;
 use sui_config::local_ip_utils;
 use sui_json_rpc_types::SuiTransactionBlockEffectsAPI;
 use sui_sdk::wallet_context::WalletContext;
@@ -48,24 +39,11 @@ use sui_test_transaction_builder::TestTransactionBuilder;
 use sui_types::base_types::ObjectRef;
 use sui_types::base_types::SequenceNumber;
 use sui_types::bridge::{BridgeChainId, TOKEN_ID_USDC};
-use sui_types::committee::TOTAL_VOTING_POWER;
 use sui_types::object::Owner;
 use sui_types::transaction::{CallArg, ObjectArg};
 use sui_types::{base_types::SuiAddress, crypto::get_key_pair, digests::TransactionDigest};
 use sui_types::{BRIDGE_PACKAGE_ID, SUI_BRIDGE_OBJECT_ID};
 use tokio::task::JoinHandle;
-use tracing::error;
-use tracing::info;
-
-pub const BRIDGE_COMMITTEE_NAME: &str = "BridgeCommittee";
-pub const SUI_BRIDGE_NAME: &str = "SuiBridge";
-pub const BRIDGE_CONFIG_NAME: &str = "BridgeConfig";
-pub const BRIDGE_LIMITER_NAME: &str = "BridgeLimiter";
-pub const BRIDGE_VAULT_NAME: &str = "BridgeVault";
-pub const BTC_NAME: &str = "BTC";
-pub const ETH_NAME: &str = "ETH";
-pub const USDC_NAME: &str = "USDC";
-pub const USDT_NAME: &str = "USDT";
 
 pub const DUMMY_MUTALBE_BRIDGE_OBJECT_ARG: ObjectArg = ObjectArg::SharedObject {
     id: SUI_BRIDGE_OBJECT_ID,
@@ -408,206 +386,4 @@ pub async fn approve_action_with_validator_secrets(
         "Didn't find the creted object owned by {}",
         expected_token_receiver
     );
-}
-
-pub async fn get_eth_signer_client_e2e_test_only(
-    eth_rpc_url: &str,
-) -> anyhow::Result<(EthSigner, String)> {
-    // This private key is derived from the default anvil setting.
-    // Mnemonic:          test test test test test test test test test test test junk
-    // Derivation path:   m/44'/60'/0'/0/
-    // DO NOT USE IT ANYWHERE ELSE EXCEPT FOR RUNNING AUTOMATIC INTEGRATION TESTING
-    let url = eth_rpc_url.to_string();
-    let private_key_0 = "0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1fcdbf7cbf4356";
-    let signer_0 = get_eth_signer_client(&url, private_key_0).await?;
-    Ok((signer_0, private_key_0.to_string()))
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Clone)]
-pub struct DeployedSolContracts {
-    pub sui_bridge: EthAddress,
-    pub bridge_committee: EthAddress,
-    pub bridge_limiter: EthAddress,
-    pub bridge_vault: EthAddress,
-    pub bridge_config: EthAddress,
-    pub btc: EthAddress,
-    pub eth: EthAddress,
-    pub usdc: EthAddress,
-    pub usdt: EthAddress,
-}
-
-impl DeployedSolContracts {
-    pub fn eth_adress_to_hex(addr: EthAddress) -> String {
-        format!("{:x}", addr)
-    }
-
-    pub fn sui_bridge_addrress_hex(&self) -> String {
-        Self::eth_adress_to_hex(self.sui_bridge)
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct SolDeployConfig {
-    committee_member_stake: Vec<u64>,
-    committee_members: Vec<String>,
-    min_committee_stake_required: u64,
-    source_chain_id: u64,
-    supported_chain_ids: Vec<u64>,
-    supported_chain_limits_in_dollars: Vec<u64>,
-    supported_tokens: Vec<String>,
-    token_prices: Vec<u64>,
-}
-
-pub async fn deploy_sol_contract(
-    anvil_url: &str,
-    eth_signer: EthSigner,
-    bridge_authority_keys: Vec<BridgeAuthorityKeyPair>,
-    tx: tokio::sync::oneshot::Sender<DeployedSolContracts>,
-    eth_private_key_hex: String,
-) {
-    let sol_path = format!("{}/../../bridge/evm", env!("CARGO_MANIFEST_DIR"));
-
-    // Write the deploy config to a temp file then provide it to the forge late
-    let deploy_config_path = tempfile::tempdir()
-        .unwrap()
-        .into_path()
-        .join("sol_deploy_config.json");
-    let node_len = bridge_authority_keys.len();
-    let stake = TOTAL_VOTING_POWER / (node_len as u64);
-    let committee_members = bridge_authority_keys
-        .iter()
-        .map(|k| {
-            format!(
-                "0x{:x}",
-                BridgeAuthorityPublicKeyBytes::from(&k.public).to_eth_address()
-            )
-        })
-        .collect::<Vec<_>>();
-    let committee_member_stake = vec![stake; node_len];
-    let deploy_config = SolDeployConfig {
-        committee_member_stake: committee_member_stake.clone(),
-        committee_members: committee_members.clone(),
-        min_committee_stake_required: 10000,
-        source_chain_id: BridgeChainId::EthCustom as u64,
-        supported_chain_ids: vec![1, 2, 3],
-        supported_chain_limits_in_dollars: vec![
-            1000000000000000,
-            1000000000000000,
-            1000000000000000,
-        ],
-        supported_tokens: vec![], // this is set up in the deploy script
-        token_prices: vec![12800, 432518900, 25969600, 10000, 10000],
-    };
-
-    let serialized_config = serde_json::to_string_pretty(&deploy_config).unwrap();
-    tracing::debug!(
-        "Serialized config written to {:?}: {:?}",
-        deploy_config_path,
-        serialized_config
-    );
-    let mut file = File::create(deploy_config_path.clone()).unwrap();
-    file.write_all(serialized_config.as_bytes()).unwrap();
-
-    // override for the deploy script
-    std::env::set_var("OVERRIDE_CONFIG_PATH", deploy_config_path.to_str().unwrap());
-    std::env::set_var("PRIVATE_KEY", eth_private_key_hex);
-    std::env::set_var("ETHERSCAN_API_KEY", "n/a");
-
-    info!("Deploying solidity contracts");
-    Command::new("forge")
-        .current_dir(sol_path.clone())
-        .arg("clean")
-        .status()
-        .expect("Failed to execute `forge clean`");
-
-    let mut child = Command::new("forge")
-        .current_dir(sol_path)
-        .arg("script")
-        .arg("script/deploy_bridge.s.sol")
-        .arg("--fork-url")
-        .arg(anvil_url)
-        .arg("--broadcast")
-        .arg("--ffi")
-        .arg("--chain")
-        .arg("31337")
-        .stdout(std::process::Stdio::piped()) // Capture stdout
-        .stderr(std::process::Stdio::piped()) // Capture stderr
-        .spawn()
-        .unwrap();
-
-    let mut stdout = child.stdout.take().expect("Failed to open stdout");
-    let mut stderr = child.stderr.take().expect("Failed to open stderr");
-
-    // Read stdout/stderr to String
-    let mut s = String::new();
-    stdout.read_to_string(&mut s).unwrap();
-    let mut e = String::new();
-    stderr.read_to_string(&mut e).unwrap();
-
-    // Wait for the child process to finish and collect its status
-    let status = child.wait().unwrap();
-    if status.success() {
-        info!("Solidity contract deployment finished successfully");
-    } else {
-        error!(
-            "Solidity contract deployment exited with code: {:?}",
-            status.code()
-        );
-    }
-    println!("Stdout: {}", s);
-    println!("Stdout: {}", e);
-
-    let mut deployed_contracts = BTreeMap::new();
-    // Process the stdout to parse contract addresses
-    for line in s.lines() {
-        if line.contains("[Deployed]") {
-            let replaced_line = line.replace("[Deployed]", "");
-            let trimmed_line = replaced_line.trim();
-            let parts: Vec<&str> = trimmed_line.split(':').collect();
-            if parts.len() == 2 {
-                let contract_name = parts[0].to_string().trim().to_string();
-                let contract_address = EthAddress::from_str(parts[1].to_string().trim()).unwrap();
-                deployed_contracts.insert(contract_name, contract_address);
-            }
-        }
-    }
-
-    let contracts = DeployedSolContracts {
-        sui_bridge: *deployed_contracts.get(SUI_BRIDGE_NAME).unwrap(),
-        bridge_committee: *deployed_contracts.get(BRIDGE_COMMITTEE_NAME).unwrap(),
-        bridge_config: *deployed_contracts.get(BRIDGE_CONFIG_NAME).unwrap(),
-        bridge_limiter: *deployed_contracts.get(BRIDGE_LIMITER_NAME).unwrap(),
-        bridge_vault: *deployed_contracts.get(BRIDGE_VAULT_NAME).unwrap(),
-        btc: *deployed_contracts.get(BTC_NAME).unwrap(),
-        eth: *deployed_contracts.get(ETH_NAME).unwrap(),
-        usdc: *deployed_contracts.get(USDC_NAME).unwrap(),
-        usdt: *deployed_contracts.get(USDT_NAME).unwrap(),
-    };
-    let eth_bridge_committee =
-        EthBridgeCommittee::new(contracts.bridge_committee, eth_signer.clone().into());
-    for (i, (m, s)) in committee_members
-        .iter()
-        .zip(committee_member_stake.iter())
-        .enumerate()
-    {
-        let eth_address = EthAddress::from_str(m).unwrap();
-        assert_eq!(
-            eth_bridge_committee
-                .committee_index(eth_address)
-                .await
-                .unwrap(),
-            i as u8
-        );
-        assert_eq!(
-            eth_bridge_committee
-                .committee_stake(eth_address)
-                .await
-                .unwrap(),
-            *s as u16
-        );
-        assert!(!eth_bridge_committee.blocklist(eth_address).await.unwrap());
-    }
-    tx.send(contracts).unwrap();
 }
